requirements.txt
gradio
opencv-python
pillow
numpy
transformers
torch
scikit-learn
